+++ src/main/java/gaji/service/domain/test/TestController.java
@@ -1,8 +1,7 @@
 package gaji.service.domain.test;
 
-import gaji.service.global.common.base.BaseResponse;
+import gaji.service.global.base.BaseResponse;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.Mapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
