--- src/main/java/gaji/service/global/common/exception/RestApiException.java
@@ -1,13 +1,0 @@
-package gaji.service.global.common.exception;
-
-import lombok.AllArgsConstructor;
-
-@AllArgsConstructor
-public class RestApiException extends RuntimeException {
-
-    private final ErrorCodeInterface errorCode;
-
-    public ErrorCode getErrorCode() {
-        return this.errorCode.getErrorCode();
-    }
-}
