+++ src/main/java/gaji/service/global/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package gaji.service.global.common.exception;
+package gaji.service.global.exception;
 
 import lombok.Builder;
 import lombok.Getter;
