+++ src/main/java/gaji/service/global/exception/ExceptionAdvice.java
@@ -1,7 +1,7 @@
-package gaji.service.global.common.exception;
+package gaji.service.global.exception;
 
-import gaji.service.global.common.base.BaseResponse;
-import gaji.service.global.common.exception.code.GlobalErrorCode;
+import gaji.service.global.base.BaseResponse;
+import gaji.service.global.exception.code.GlobalErrorCode;
 import lombok.extern.slf4j.Slf4j;
 import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.http.HttpHeaders;
